fix: Minor refactor while doing PR
